requirements.txt
pandas
seaborn
matplotlib
scikit-learn
numpy
catboost
xgboost
numpy
Flask
python-dotenv
pymongo
certifi
dill
pyaml

# -e .